% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model_xy.R
\name{copulareg.default}
\alias{copulareg.default}
\alias{copulareg}
\title{copulareg}
\usage{
copulareg.default(
  y,
  x,
  var_type_y,
  var_type_x,
  family_set = c("gaussian", "clayton", "gumbel"),
  extra_x = NULL,
  extra_y = NULL
)
}
\arguments{
\item{y}{A vector of n observations of the (univariate) outcome variable y}

\item{x}{A (n x p) matrix of n observations of p covariates}

\item{var_type_y}{A character that has to be specified as "d" or "c"
to indicate whether y is discrete or continuous, respectively.}

\item{var_type_x}{A vector of p characters that have to take the value
"c" or "d" to indicate whether each margin of the covariates is discrete
or continuous.}

\item{family_set}{A vector of strings that specifies the set of pair-copula
families that the fitting algorithm chooses from. For a overview of which
values that can be specified, see the documentation for \link[rvinecopulib]{bicop}.}

\item{extra_x}{Optional extra values of x to use for estimating the margins
of the covariates.}

\item{extra_y}{Optional extra values of y to use to estimate the margin of y.}
}
\description{
This function fits joint distributions with an R-vine
pair copula structure, that is constructed in a specific way so that
the conditional density and distribution of the variable y can be
computed explicitly.
}
